% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_variants.R
\name{flag_variants}
\alias{flag_variants}
\title{Flag and annotate variants using external databases and panel of normals information}
\usage{
flag_variants(variants, normal_variants, flag_patterns, exon_ranges,
  homop_ranges, RNAedit_ranges, repeats_ranges, flag_qualities = TRUE)
}
\arguments{
\item{variants}{dataframe containing all the variants in the cohort. This is created within the `variants_power()` function}

\item{normal_variants}{link to panel of normal file parsed with `samplepower::parse_pon`}

\item{flag_patterns}{dataframe containing the flags to assign to variants. THIS NEEDS TO BE SET AS INTERNAL TO THE PACKAGE}

\item{exon_ranges}{GRanges object with exon boundaries relative to the reference genome used}

\item{homop_ranges}{GRanges object with homopolymers regions of more than 5bp relative to the reference genome used}

\item{RNAedit_ranges}{GRanges object with RNA editing sites relative to the reference genome used}

\item{repeats_ranges}{GRanges object with repetitive regions of the reference genome used}

\item{flag_qualities}{logical. TRUE by default, variants are flagged based on quality measures. If FALSE this step is skipped.}
}
\description{
Flag and annotate variants using external databases and panel of normals information
}
