% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gatk_coverage.R
\name{parse_gatk_coverage}
\alias{parse_gatk_coverage}
\title{Parse and coverage and VAF from every location in the truth to the downsampled variants provided}
\usage{
parse_gatk_coverage(path_to_gatk_coverage, truth_set, TCGA = FALSE)
}
\arguments{
\item{path_to_gatk_coverage}{path to file where the GATK `DepthOfCoverage` output is saved}

\item{truth_set}{external truth set}

\item{TCGA}{logical. If TRUE variants will be macthed with truth sets excluding transcript information}
}
\description{
Parse and coverage and VAF from every location in the truth to the downsampled variants provided
}
