% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pon.R
\name{parse_pon}
\alias{parse_pon}
\title{Parse panel of normal variants to standardise output across callers}
\usage{
parse_pon(path_to_pon, caller)
}
\arguments{
\item{path_to_pon}{link to panel of normal file created with GATK3 VariantsToTable using at least}

\item{caller}{one of `varscan`, `mutect` or `vardict`. This field is needed to parse the variant allele frequency field which is recorded differently across callers.}
}
\description{
Parse panel of normal variants to standardise output across callers
}
\examples{

path_to_pon = "varscan_calls/PON_varscan_target_regions.table"

parse_varscan <- parse_pon(path_to_pon,'varscan')
}
