% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_names.R
\name{reassign_names}
\alias{reassign_names}
\title{Rename columns of `featureCounts` counts output}
\usage{
reassign_names(sample_names, counts)
}
\arguments{
\item{counts}{matrix of counts from `featureCounts`}

\item{sampleNames}{character with sample name}
}
\description{
Rename columns of `featureCounts` counts output
}
