% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_compute.R
\name{sens_compute}
\alias{sens_compute}
\title{Computes sensitivity with respect to a truth set}
\usage{
sens_compute(variants, truth_set, indel = FALSE)
}
\arguments{
\item{variants}{dataframe created within the `variant_power.R` function. The columns required in this dataframe are `Called_defaults` and `Called_defaults` which are numeric 0/1 defining whther a variant is called. If `indel = TRUE` also the column `key` is required. `key` is a unique key for a variant: `chr:POS:SampleName:Gene:Transcript`.}

\item{truth_set}{external truth set. The truth set os only needed to know the total number of variants in it.}

\item{indel}{logical. TRUE is INDELs are analysed.}
}
\description{
Computes sensitivity with respect to a truth set
}
