% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_by_threshold.R
\name{sensitivity_by_thresholds}
\alias{sensitivity_by_thresholds}
\title{This function is a wrapper for `sens_compute()` and computes sensitivity for varying VAF and total depth.}
\usage{
sensitivity_by_thresholds(variants, truth_set, indel = FALSE)
}
\arguments{
\item{variants}{dataframe created within the `variant_power.R` function. The columns required in this dataframe are `Called_defaults` and `Called_defaults` which are numeric 0/1 defining whther a variant is called. If `indel = TRUE` also the column `key` is required. `key` is a unique key for a variant: `chr:POS:SampleName:Gene:Transcript`.}

\item{truth_set}{external truth set}

\item{indel}{logical. TRUE is INDELs are analysed.}
}
\description{
This function is a wrapper for `sens_compute()` and computes sensitivity for varying VAF and total depth.
}
