% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants_power.R
\name{variants_power}
\alias{variants_power}
\title{Compute sensitivity and specificity for the a set of variants with respect to some truth sets}
\usage{
variants_power(variant_files, variant_files_initial, down_label = NA,
  truth_set = NA, caller, gene_expression = NA, normal_variants,
  exon_ranges, homop_ranges, RNAedit_ranges, repeats_ranges, ncbi,
  flag_patterns, path_to_gatk_coverage, TCGA = FALSE)
}
\arguments{
\item{variant_files}{character vector containing the paths to all the tab delimited files of variants from the downsampled run. See details.}

\item{variant_files_initial}{character vector containing the paths to all the tab delimited files of variants from the initial run with deeply sequenced RNA-Seq samples. See details.}

\item{down_label}{character. Any label to be assigned to the current run}

\item{truth_set}{external truth set}

\item{caller}{one of `varscan`, `mutect` or `vardict`}

\item{gene_expression}{if available, `rds` objects with raw gene expression data where rows are genes and columns are samples}

\item{normal_variants}{link to panel of normal file parsed with `samplepower::parse_pon`}

\item{exon_ranges}{GRanges object with exon boundaries relative to the reference genome used}

\item{homop_ranges}{GRanges object with homopolymers regions of more than 5bp relative to the reference genome used}

\item{RNAedit_ranges}{GRanges object with RNA editing sites relative to the reference genome used}

\item{repeats_ranges}{GRanges object with repetitive regions of the reference genome used}

\item{ncbi}{NCBI annotation dataframe including at least the columns `GeneID` and `Symbol`}

\item{flag_patterns}{dataframe containing the flags to assign to variants. THIS NEEDS TO BE SET AS INTERNAL TO THE PACKAGE}

\item{path_to_gatk_coverage}{path to file where the GATK `DepthOfCoverage` output is saved}

\item{TCGA}{logical. If TRUE variants will be macthed with truth sets excluding transcript information}
}
\description{
Compute sensitivity and specificity for the a set of variants with respect to some truth sets
}
\details{
The files whose links are recorded in `variant_files_initial` and `variant_files` are the ouput files obtained with the `call_variants.R` function from the functions in https://github.com/annaquaglieri16/RNA-seq-variant-calling#panel-of-normals-pon. The VCF output from every callers are standardised to make the results easily comparable across callers.
}
